backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.52.0
pydantic==2.11.1
python-dotenv==1.1.0
supabase==2.13.0
